# specs/007-global-chatbot/contracts/api-contract.yaml
openapi: 3.0.3
info:
  title: Chatbot API
  description: API for the Global Chatbot UI Integration with Docusaurus Frontend
  version: 1.0.0
  contact:
    name: Physical AI & Humanoid Robotics Support
    email: support@physical-ai-book.com

servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.physical-ai-book.com/api/v1
    description: Production server

paths:
  /chat:
    post:
      summary: Submit a query to the chatbot
      description: Sends a user query to the backend AI service and returns a response
      operationId: chat_endpoint
      tags:
        - chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            example:
              query: "What are the key concepts in chapter 3?"
              selected_text: "Chapter 3 discusses neural networks and deep learning architectures"
              restrict_to_selection: false
              conversation_id: "conv_12345"
      responses:
        "200":
          description: Successful response from the chatbot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatSuccessResponse"
              example:
                success: true
                data:
                  response: "Chapter 3 covers neural networks and deep learning architectures, including convolutional neural networks and recurrent neural networks."
                  sources: ["chapter_3_neural_networks.pdf", "deep_learning_architectures.md"]
                  confidence: 0.95
                  conversation_id: "conv_12345"
                  timestamp: "2026-01-21T10:30:00Z"
                request_id: "req_abc123"
        "400":
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatErrorResponse"
              example:
                success: false
                error:
                  code: "INVALID_QUERY"
                  message: "Query must be provided and cannot be empty"
                  details:
                    field: "query"
                    reason: "Field is required and cannot be empty"
                request_id: "req_def456"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatErrorResponse"
              example:
                success: false
                error:
                  code: "INTERNAL_ERROR"
                  message: "An internal error occurred while processing the request"
                  details: {}
                request_id: "req_ghi789"

components:
  schemas:
    ChatRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: The user's query or question
          minLength: 1
          maxLength: 2000
          example: "Explain the concept of reinforcement learning"
        selected_text:
          type: string
          description: Optional text selected by the user to provide context
          maxLength: 5000
          example: "Reinforcement learning is a type of machine learning where an agent learns to make decisions by performing actions in an environment."
        restrict_to_selection:
          type: boolean
          description: Whether to limit the response to the selected text only
          default: false
        conversation_id:
          type: string
          description: Optional ID to maintain conversation context
          pattern: "^[a-zA-Z0-9_-]+$"
          example: "conv_12345"

    ChatResponse:
      type: object
      required:
        - response
        - conversation_id
        - timestamp
      properties:
        response:
          type: string
          description: The chatbot's response to the user's query
          example: "Reinforcement learning is a type of machine learning where an agent learns to make decisions by performing actions in an environment..."
        sources:
          type: array
          description: List of sources used to generate the response
          items:
            type: string
          maxItems: 10
          example: ["reinforcement_learning_chapter.pdf", "rl_examples.md"]
        confidence:
          type: number
          description: Confidence level of the response (0-1)
          minimum: 0
          maximum: 1
          example: 0.85
        conversation_id:
          type: string
          description: ID of the conversation
          example: "conv_12345"
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the response was generated
          example: "2026-01-21T10:30:00Z"

    ChatSuccessResponse:
      type: object
      required:
        - success
        - data
        - request_id
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
          example: true
        data:
          $ref: "#/components/schemas/ChatResponse"
        request_id:
          type: string
          description: Unique ID for the request for tracking purposes
          example: "req_abc123"

    ChatErrorResponse:
      type: object
      required:
        - success
        - error
        - request_id
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
          example: false
        error:
          type: object
          required:
            - code
            - message
            - details
          properties:
            code:
              type: string
              description: Error code
              example: "INVALID_QUERY"
            message:
              type: string
              description: Human-readable error message
              example: "Query must be provided and cannot be empty"
            details:
              type: object
              description: Additional error details
              example: {"field": "query", "reason": "Field is required and cannot be empty"}
        request_id:
          type: string
          description: Unique ID for the request for tracking purposes
          example: "req_def456"

tags:
  - name: chat
    description: Chatbot interaction endpoints